# CORE-1390 Change underlying data strucure to atomic reference.