## Overview
before
```
[info] Benchmark                      Mode  Cnt     Score      Error  Units
[info] WideBench.inmemWideReduceFine    ss    5   342.613 ±  142.595  ms/op
[info] WideBench.wideReduceFine         ss    5  3127.114 ± 1771.273  ms/op
```

after
```
[info] Benchmark                      Mode  Cnt     Score     Error  Units
[info] WideBench.inmemWideReduceFine    ss    5   214.313 ±  22.963  ms/op
[info] WideBench.wideReduceFine         ss    5  2459.418 ± 357.525  ms/op
```
monix
```
[info] Benchmark                      Mode  Cnt     Score     Error  Units
[info] WideBench.inmemWideReduceFine    ss    5   162.638 ± 124.215  ms/op
[info] WideBench.wideReduceFine         ss    5  2465.210 ± 532.292  ms/op
```


### Which JIRA issue does this PR relate to? If there is not a JIRA issue addressing this work, please create one now and add the link here.
https://rchain.atlassian.net/projects/CORE/issues/CORE-1390

### Complete this checklist before you submit the PR
- [x] This PR contains no more than 200 lines of code, excluding test code.
- [x] This PR meets [RChain development coding standards](https://rchain.atlassian.net/wiki/spaces/DOC/pages/28082177/Coding+Standards).
- [ ] If this PR adds a new feature, this PR includes tests related to this feature.
- [x] You assigned one person to review this PR

### If you are not a member of the core development team, please confirm:
- [x] You signed the commit. Merging requires a signature. Please see the [developer wiki](https://rchain.atlassian.net/wiki/spaces/DOC/pages/498630673/How+to+sign+commits+to+rchain+rchain) for instructions.
- [x] Your GitHub account is also an account with [Travis CI](https://travis-ci.org). Unit tests will not run on your PR unless you have an account with Travis. Merge requires passed unit tests.

### Notes
Optional. Add any notes on caveats, approaches you tried that didn't work, or anything else.