@lukasz-golebiewski 2018-10-30
Let's go ahead with `TrieMap` as discussed offline